pousada 0.0.15 - padronização

+++ Backend/clientes2/src/main/java/org/example/clientes/model/repostory/QuartoRepository.java
@@ -3,5 +3,5 @@ package org.example.clientes.model.repostory;
 import org.example.clientes.model.entity.Quarto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuartoRepository extends JpaRepository<Quarto, Integer> {
+public interface QuartoRepository extends JpaRepository<Quarto, Long> {
 }
